+++ src/JavLuv/MainWindowViewModel.cs
@@ -351,7 +351,9 @@ namespace JavLuv
             if (JavLuv.Settings.Get().EnableMoveRename && JavLuv.Settings.Get().MoveRenameAfterScan && m_movieScanner.IsCancelled == false)
                 m_movieBrowserViewModel.MoveRenameMovies(m_movieScanner.Movies);
 
+            // Add new actresses
             m_movieCollection.AddMovies(m_movieScanner.Movies);
+
             m_movieScanner.Clear();
         }
 
@@ -363,23 +365,18 @@ namespace JavLuv
                 PercentComplete = 0;
                 ScanStatus = string.Format(TextManager.GetString("Text.ScanningFolders"), m_movieScanner.ItemsProcessed);
             }
-            else if (m_movieScanner.Phase == ScanPhase.LoadingMetadata)
+            else
             {
                 ProgressState = TaskbarItemProgressState.Normal;
                 PercentComplete = (int)(((float)m_movieScanner.ItemsProcessed / (float)m_movieScanner.TotalItems) * 100.0);
-                ScanStatus = string.Format(TextManager.GetString("Text.LoadingMetadata"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
-            }
-            else if (m_movieScanner.Phase == ScanPhase.DownloadMetadata)
-            {
-                ProgressState = TaskbarItemProgressState.Normal;
-                PercentComplete = (int)(((float)m_movieScanner.ItemsProcessed / (float)m_movieScanner.TotalItems) * 100.0);
-                ScanStatus = string.Format(TextManager.GetString("Text.DownloadingMetadata"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
-            }
-            else if (m_movieScanner.Phase == ScanPhase.DownloadActressData)
-            {
-                ProgressState = TaskbarItemProgressState.Normal;
-                PercentComplete = (int)(((float)m_movieScanner.ItemsProcessed / (float)m_movieScanner.TotalItems) * 100.0);
-                ScanStatus = string.Format(TextManager.GetString("Text.DownloadingActressData"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
+                if (m_movieScanner.Phase == ScanPhase.LoadingMetadata)
+                    ScanStatus = string.Format(TextManager.GetString("Text.LoadingMetadata"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
+                else if (m_movieScanner.Phase == ScanPhase.ImportMovies)
+                    ScanStatus = string.Format(TextManager.GetString("Text.ImportingMovies"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
+                else if (m_movieScanner.Phase == ScanPhase.DownloadMetadata)
+                    ScanStatus = string.Format(TextManager.GetString("Text.DownloadingMetadata"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
+                else if (m_movieScanner.Phase == ScanPhase.DownloadActressData)
+                    ScanStatus = string.Format(TextManager.GetString("Text.DownloadingActressData"), m_movieScanner.ItemsProcessed, m_movieScanner.TotalItems);
             }
         }
 
@@ -481,7 +478,7 @@ namespace JavLuv
             NotifyPropertyChanged("ScanVisibility");
         }
 
-        public void StartScan(List<string> scanDirectories)
+        public void StartRecan(List<string> scanDirectories)
         {
             m_movieScanner.Start(scanDirectories);
             SidePanel.SettingsIsEnabled = false;